Move `status-report.js` out of `runtime-test`
